Fix `systemScreenMessage()` for scale and Unicode
Fix `systemScreenMessage()` for wxWidgets (the OSD messages) to be the
same scale regardless of filter scale and to display Unicode correctly.

There is enough space for 3 lines at the bottom, which should be
generally sufficient.

The `drawText()` function is replaced with `drawTextWx()` for the
wxWidgets interface for this purpose, using wxWidgets facilities for
drawing text on an image.

Fixed with Claude and lots of testing.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>