build: fixup MSYS2 CLANG64 builds
Fix builds on MSYS2 CLANG64, both for MSYS2 MinGW packages and for
vcpkg.

Rename `MSYS` variable to `MSYS2`, CMake uses `MSYS` for something else.

Use pkg-config on Windows as well, both MSYS2 and vcpkg have it now.

Fix linking the MSYS2 MinGW FFmpeg statically.

Copy `FindOpenAL.cmake` to our `cmake/` directory and change the order
of library to prefer `OpenAL32` which is the vcpkg name.

Add automatic `ENV{PATH}` setting for clang64 when using the
`x64-mingw-static` triplet for vcpkg packages.

Unset `PKG_CONFIG` environment variables when using vcpkg, they are set
by MSYS2 and interfere with the vcpkg `pkg-config`.

Turn off LTO for CLANG64 as well, as it no longer works. Both in an
MSYS2 shell and when using the toolchain.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>